move RPC submitWorkDetail from namespace eth_ to parity_
